User Interface - Frequency Control
  * Operating refinements
  * Standardization across fl* apps

+++ src/include/icom/IC703.h
@@ -15,7 +15,7 @@
 // GNU General Public License for more details.
 //
 // You should have received a copy of the GNU General Public License
-// aunsigned long int with this program.  If not, see <http://www.gnu.org/licenses/>.
+// along with this program.  If not, see <http://www.gnu.org/licenses/>.
 // ----------------------------------------------------------------------------
 
 #ifndef _IC703_H
@@ -32,8 +32,8 @@ public:
 
 	bool check();
 
-	unsigned long int get_vfoA(void);
-	void set_vfoA(unsigned long int);
+	unsigned long long get_vfoA(void);
+	void set_vfoA(unsigned long long);
 
 	void set_modeA(int val);
 	int  get_modeA();
