+++ src/include/icom/IC705.h
@@ -17,7 +17,7 @@
 // GNU General Public License for more details.
 //
 // You should have received a copy of the GNU General Public License
-// aunsigned long int with this program.  If not, see <http://www.gnu.org/licenses/>.
+// along with this program.  If not, see <http://www.gnu.org/licenses/>.
 // ----------------------------------------------------------------------------
 
 #ifndef _IC705_H
@@ -39,11 +39,11 @@ public:
 
 	bool check();
 
-	unsigned long int get_vfoA ();
-	void set_vfoA (unsigned long int freq);
+	unsigned long long get_vfoA ();
+	void set_vfoA (unsigned long long freq);
 
-	unsigned long int get_vfoB(void);
-	void set_vfoB(unsigned long int f);
+	unsigned long long get_vfoB(void);
+	void set_vfoB(unsigned long long f);
 
 	void set_modeA(int val);
 	int  get_modeA();
