+++ src/include/other/AOR5K.h
@@ -15,7 +15,7 @@
 // GNU General Public License for more details.
 //
 // You should have received a copy of the GNU General Public License
-// aunsigned long int with this program.  If not, see <http://www.gnu.org/licenses/>.
+// along with this program.  If not, see <http://www.gnu.org/licenses/>.
 // ----------------------------------------------------------------------------
 
 #ifndef AOR5K_H
@@ -32,10 +32,10 @@ public:
 
 	bool check();
 
-	unsigned long int get_vfoA();
-	void set_vfoA(unsigned long int);
-	unsigned long int get_vfoB();
-	void set_vfoB(unsigned long int);
+	unsigned long long get_vfoA();
+	void set_vfoA(unsigned long long);
+	unsigned long long get_vfoB();
+	void set_vfoB(unsigned long long);
 	void set_volume_control(int val);
 	int  get_volume_control();
 	void get_vol_min_max_step(int &min, int &max, int &step);
