+++ src/include/yaesu/FT991A.h
@@ -15,7 +15,7 @@
 // GNU General Public License for more details.
 //
 // You should have received a copy of the GNU General Public License
-// aunsigned long int with this program.  If not, see <http://www.gnu.org/licenses/>.
+// along with this program.  If not, see <http://www.gnu.org/licenses/>.
 // ----------------------------------------------------------------------------
 
 #ifndef _FT991A_H
@@ -36,11 +36,11 @@ public:
 
 	bool check();
 
-	unsigned long int get_vfoA();
-	void set_vfoA(unsigned long int);
+	unsigned long long get_vfoA();
+	void set_vfoA(unsigned long long);
 
-	unsigned long int get_vfoB();
-	void set_vfoB(unsigned long int);
+	unsigned long long get_vfoB();
+	void set_vfoB(unsigned long long);
 
 	bool twovfos() { return true; }
 	bool canswap() { return true; }
