+++ src/include/yaesu/FTdx101D.h
@@ -15,7 +15,7 @@
 // GNU General Public License for more details.
 //
 // You should have received a copy of the GNU General Public License
-// aunsigned long int with this program.  If not, see <http://www.gnu.org/licenses/>.
+// along with this program.  If not, see <http://www.gnu.org/licenses/>.
 // ----------------------------------------------------------------------------
 
 #ifndef _FTdx101D_H
@@ -36,11 +36,11 @@ public:
 
 	virtual bool check();
 
-	virtual unsigned long int get_vfoA();
-	virtual void set_vfoA(unsigned long int);
+	virtual unsigned long long get_vfoA();
+	virtual void set_vfoA(unsigned long long);
 
-	virtual unsigned long int get_vfoB();
-	virtual void set_vfoB(unsigned long int);
+	virtual unsigned long long get_vfoB();
+	virtual void set_vfoB(unsigned long long);
 
 	virtual bool twovfos();
 	virtual void selectA();
