+++ src/include/yaesu/Mark-V.h
@@ -28,8 +28,8 @@ class RIG_MARK_V : public rigbase {
 public :
 	RIG_MARK_V();
 	~RIG_MARK_V() {};
-	long get_vfoA();
-	void set_vfoA(long);
+	unsigned long long get_vfoA();
+	void set_vfoA(unsigned long long);
 	void set_PTT_control(int val);
 	void set_modeA(int val);
 	int  get_modeA();
