+++ src/rigs/icom/IC756PRO2.cxx
@@ -15,7 +15,7 @@
 // GNU General Public License for more details.
 //
 // You should have received a copy of the GNU General Public License
-// aunsigned long int with this program.  If not, see <http://www.gnu.org/licenses/>.
+// along with this program.  If not, see <http://www.gnu.org/licenses/>.
 // ----------------------------------------------------------------------------
 
 #include "icom/IC756PRO2.h"
@@ -116,11 +116,14 @@ RIG_IC756PRO2::RIG_IC756PRO2() {
 	bandwidths_ = IC756PRO2_SSBwidths;
 	bw_vals_ = IC756PRO2_bw_vals_SSB;
 
+//	serial_write_delay = 0;
+//	serial_post_write_delay = 0;
+
 	_mode_type = IC756PRO2_mode_type;
 
 	widgets = IC756PRO2_widgets;
 
-	def_freq = freqA = freqB = A.freq = 14070000;
+	def_freq = freqA = freqB = A.freq = 14070000ULL;
 	def_mode = modeA = modeB = B.imode = 1;
 	def_bw = bwA = bwB = A.iBW = B.iBW = 32;
 
@@ -215,7 +218,7 @@ bool RIG_IC756PRO2::check ()
 	return ok;
 }
 
-unsigned long int RIG_IC756PRO2::get_vfoA ()
+unsigned long long RIG_IC756PRO2::get_vfoA ()
 {
 	if (inuse == onB) return A.freq;
 	std::string cstr = "\x03";
@@ -237,7 +240,7 @@ unsigned long int RIG_IC756PRO2::get_vfoA ()
 	return A.freq;
 }
 
-void RIG_IC756PRO2::set_vfoA (unsigned long int freq)
+void RIG_IC756PRO2::set_vfoA (unsigned long long freq)
 {
 	A.freq = freq;
 	cmd = pre_to;
@@ -248,7 +251,7 @@ void RIG_IC756PRO2::set_vfoA (unsigned long int freq)
 	set_trace(2, "set_vfoA()", str2hex(cmd.c_str(), cmd.length()));
 }
 
-unsigned long int RIG_IC756PRO2::get_vfoB ()
+unsigned long long RIG_IC756PRO2::get_vfoB ()
 {
 	if (inuse == onA) return B.freq;
 	std::string cstr = "\x03";
@@ -270,7 +273,7 @@ unsigned long int RIG_IC756PRO2::get_vfoB ()
 	return B.freq;
 }
 
-void RIG_IC756PRO2::set_vfoB (unsigned long int freq)
+void RIG_IC756PRO2::set_vfoB (unsigned long long freq)
 {
 	B.freq = freq;
 	cmd = pre_to;
@@ -1193,7 +1196,7 @@ void RIG_IC756PRO2::get_band_selection(int v)
 		set_trace(2, "get band stack", str2hex(replystr.c_str(), replystr.length()));
 		size_t p = replystr.rfind(pre_fm);
 		if (p != std::string::npos) {
-			unsigned long int bandfreq = fm_bcd_be(replystr.substr(p+8, 5), 10);
+			unsigned long long bandfreq = fm_bcd_be(replystr.substr(p+8, 5), 10);
 			int bandmode = replystr[p+13];
 			int bandfilter = replystr[p+14];
 			int banddata = replystr[p+15] & 0x10;
@@ -1223,7 +1226,7 @@ void RIG_IC756PRO2::get_band_selection(int v)
 
 void RIG_IC756PRO2::set_band_selection(int v)
 {
-	unsigned long int freq = (inuse == onB ? B.freq : A.freq);
+	unsigned long long freq = (inuse == onB ? B.freq : A.freq);
 	int fil = (inuse == onB ? filB : filA);
 	int mode = (inuse == onB ? B.imode : A.imode);
